backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
pydantic==2.11.7
pydantic-settings==2.11.0
python-multipart==0.0.20
Pillow==12.1.1
pytest==8.4.1
httpx==0.28.1
